simple-rag-document-api/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pydantic
tiktoken

# ── Testing ────────────────────────────────────
pytest
pytest-asyncio
httpx
